This hardens `iree/async` ownership and platform lifecycle contracts
needed by long-lived network transports. Remoting is the first
production consumer, but this pull request contains only generic async
infrastructure: retained progress ownership, explicit socket setup
state, direct IOCP event sources, and io_uring resource management that
respects the kernel's single-issuer task identity.

## Ownership and progress

- Buffer pool handles are reference counted, and leases are explicitly
move-only, so registered storage can retain its allocation mechanism
through asynchronous completion and teardown.
- Proactor-pool entries independently retain both the proactor and its
runner. Releasing the aggregate pool no longer stops progress for
retained consumers, while aggregate shutdown still requests every owned
runner to stop before joining any of them.
- Pool entries remain lazily created, borrowed lookup remains available
for pool-owned consumers, and queue submission and completion paths gain
no ownership atomics.

## Socket and event lifecycle

- Socket binding is represented as explicit setup state: created sockets
begin unbound, accepted sockets begin bound, imported handles begin
unknown, failed binds restore their prior state, and duplicate or
concurrent binds fail before entering platform APIs.
- Windows event sources use NT wait completion packets to deliver waits
directly through IOCP. Unsupported systems fail registration instead of
silently adding a threadpool-backed progress path.
- The async and HAL CTS entry points now initialize tracing for the
complete test lifetime.

## io_uring ownership

- Event sources and relays register as logical pending monitors. The
poll owner materializes them into SQEs in ring-sized batches, allowing
monitor populations larger than the submission queue while preserving
`IORING_SETUP_SINGLE_ISSUER`.
- Every ring registration opcode passes through a package-private owner
dispatcher. Calls before owner binding or from the owner execute
directly; other callers queue a stack-backed synchronous request, wake
the existing poll loop, and wait without an arbitrary deadline. Normal
submission, ring entry, and completion paths gain no mutex.
- Permanent poll retirement closes the dispatcher and releases queued
callers with `ESHUTDOWN`, so resource teardown cannot wait on an owner
that no longer exists.
- Zero-copy socket sends retain the primary completion result until the
ownership notification, preserving the real status and byte count while
deferring payload release until the kernel is done.

## Registered buffers

- Fixed buffers, provided-buffer rings, and probes all use the
poll-owner registration path. Sparse-table bookkeeping never remains
locked across the owner handoff, and partial kernel updates are rolled
back exactly.
- Legacy fixed-buffer registration has an explicit lifecycle,
provided-buffer groups search the full 16-bit identifier space with the
kernel as collision authority, and teardown does not free userspace
metadata while a kernel reference may remain.
- DMA-buffer registration requests the writable long-term pin required
by io_uring while preserving the caller's published access. Memory that
cannot satisfy that optimization remains usable through the mapped copy
path; unexpected kernel failures remain hard errors.

+++ runtime/src/iree/async/region.h
@@ -117,7 +117,7 @@ enum iree_async_region_type_e {
   // buffer ring (PBUF_RING) with a buffer_group_id for kernel-managed buffer
   // selection during recv. For READ-access slabs, registers into the kernel's
   // fixed buffer table with a base_buffer_index for zero-copy send via
-  // IORING_OP_SEND_ZC with IOSQE_FIXED_FILE.
+  // IORING_OP_SEND_ZC with IORING_RECVSEND_FIXED_BUF.
   IREE_ASYNC_REGION_TYPE_IOURING,
 
   // DMA buffer backed by a file descriptor. Used for GPU-accessible memory
@@ -179,27 +179,37 @@ typedef struct iree_async_region_t {
   // Backend-specific handles produced by registration.
   union {
     struct {
-      uint32_t lkey;  // Local access key.
-      uint32_t rkey;  // Remote access key.
-      void* mr;       // ibv_mr* or equivalent.
+      // Local access key.
+      uint32_t lkey;
+
+      // Remote access key.
+      uint32_t rkey;
+
+      // ibv_mr* or equivalent backend memory-region handle.
+      void* mr;
     } rdma;
     struct {
       // For provided buffer ring (recv - kernel selects buffer).
       // -1 if not in a provided buffer ring.
-      int16_t buffer_group_id;
+      int32_t buffer_group_id;
 
       // For fixed buffer table (send - application selects buffer).
       // buf_index = span.offset / buffer_size + base_buffer_index
       // Starting index in kernel's fixed buffer table.
       // -1 if not registered in the kernel's fixed buffer table (e.g.,
       // RLIMIT_MEMLOCK too low to pin pages); send falls back to copy I/O.
-      int16_t base_buffer_index;
+      int32_t base_buffer_index;
     } iouring;
     struct {
+      // DMA buffer file descriptor. Not owned.
       int fd;
+
+      // Byte offset of this region in the DMA buffer.
       uint64_t offset;
     } dmabuf;
-    uint64_t opaque[4];  // Catch-all for unknown backends.
+
+    // Catch-all storage for backend handles without a public representation.
+    uint64_t opaque[4];
   } handles;
 } iree_async_region_t;
